telegrambot/requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-telegram-bot==13.3
pytz==2021.1
six==1.15.0
tornado==6.1
tzlocal==2.1
Werkzeug==1.0.1
mysql-connector-python==8.0.25
